Complete isolation of validation code
