+++ sigoptlite/validators.py
@@ -1,4 +1,5 @@
 import numpy
+import itertools
 
 from libsigopt.aux.geometry_utils import find_interior_point
 
@@ -8,6 +9,8 @@ from libsigopt.aux.constant import (
   ConstraintType,
   ParameterTransformationNames,
 )
+
+from sigoptlite.models import parameter_conditions_satisfied
 
 EXPERIMENT_CREATE_SCHEMA = {
   "definitions": {
@@ -186,7 +189,153 @@ OBSERVATION_CREATE_SCHEMA = {
 def get_num_distinct_elements(lst):
     return len(set(lst))
 
-def check_all_conditional_values_satisfied(experiment):
+def validate_experiment(experiment, cls_name):
+    validate_parameters_for_experiment(experiment)
+    validate_metrics_for_experiment(experiment)
+
+    if not experiment.parallel_bandwidth == 1:
+        raise ValueError(f"{cls_name} must have parallel_bandwidth == 1")
+
+    observation_budget = experiment.observation_budget
+    if observation_budget is None:
+        if experiment.num_solutions > 1:
+            raise ValueError(f"observation_budget is required for a {cls_name} with multiple solutions")
+        if experiment.requires_pareto_frontier_optimization:
+            raise ValueError(f"observation_budget is required for a {cls_name} with more than one optimized metric")
+        if experiment.has_constraint_metrics:
+            raise ValueError(f"observation_budget is required for a {cls_name} with constraint metrics")
+        if experiment.is_multitask:
+            raise ValueError(f"observation_budget is required for a {cls_name} with tasks (multitask)")
+
+    if not (experiment.optimized_metrics or experiment.constraint_metrics):
+        raise ValueError(f"{cls_name} must have optimized or constraint metrics")
+
+    if experiment.optimized_metrics:
+        if not len(experiment.optimized_metrics) in [1, 2]:
+            raise ValueError(f"{cls_name} must have one or two optimized metrics")
+        elif len(experiment.optimized_metrics) == 1 and experiment.optimized_metrics[0].threshold is not None:
+            raise ValueError(
+              "Thresholds are only supported for experiments with more than one optimized metric."
+              " Try an All-Constraint experiment instead by setting `strategy` to `constraint`."
+            )
+
+    # Check feature viability of multisolution experiments
+    num_solutions = experiment.num_solutions
+    if num_solutions and num_solutions > 1:
+        if num_solutions > observation_budget:
+            raise ValueError("observation_budget needs to be larger than the number of solutions")
+        if not len(experiment.optimized_metrics) == 1:
+            raise ValueError(f"{cls_name} with multiple solutions require exactly one optimized metric")
+
+    # Check conditional limitation
+    parameters_have_conditions = any(parameter.conditions for parameter in experiment.parameters)
+    if parameters_have_conditions ^ experiment.is_conditional:
+        raise ValueError(
+          f"For conditional {cls_name}, need both conditions defined in parameters and conditionals variables"
+          " defined in experiment"
+        )
+    if experiment.is_conditional:
+        if num_solutions and num_solutions > 1:
+            raise ValueError(f"{cls_name} with multiple solutions does not support conditional parameters")
+        if experiment.is_search:
+            raise ValueError(f"All-Constraint {cls_name} does not support conditional parameters")
+        validate_conditionals_for_experiment(experiment)
+
+    # Check feature viability of multitask
+    tasks = experiment.tasks
+    if tasks:
+        if experiment.requires_pareto_frontier_optimization:
+            raise ValueError(f"{cls_name} cannot have both tasks and multiple optimized metrics")
+        if experiment.has_constraint_metrics:
+            raise ValueError(f"{cls_name} cannot have both tasks and constraint metrics")
+        if num_solutions and num_solutions > 1:
+            raise ValueError(f"{cls_name} with multiple solutions cannot be multitask")
+        validate_tasks_for_experiment(experiment, cls_name)
+
+    if experiment.linear_constraints:
+        validate_constraints_for_experiment(experiment)
+        check_constraint_feasibility_for_experiment(experiment)
+
+def validate_parameter(parameter):
+    # categorical parameter
+    if parameter.is_categorical:
+        if not len(parameter.categorical_values) > 1:
+            raise ValueError(
+              f"Categorical parameter {parameter.name} must have more than one categorical value. "
+              f"Current values are {parameter.categorical_values}"
+            )
+        if parameter.grid:
+            raise ValueError("Categorical parameter does not support grid values")
+        if parameter.bounds:
+            raise ValueError(f"Categorical parameter should not have bounds: {parameter.bounds}")
+
+    # parameter with grid
+    if parameter.grid:
+        if not len(parameter.grid) > 1:
+            raise ValueError(
+              f"Grid parameter {parameter.name} must have more than one value. Current values are {parameter.grid}"
+            )
+        if parameter.bounds:
+            raise ValueError(f"Grid parameter should not have bounds: {parameter.bounds}")
+        if not get_num_distinct_elements(parameter.grid) == len(parameter.grid):
+            raise ValueError(f"Grid values should be unique: {parameter.grid}")
+
+    # log transformation
+    if parameter.has_transformation:
+        if not parameter.is_double:
+            raise ValueError("Transformation only applies to parameters type of double")
+        if parameter.bounds and parameter.bounds.min <= 0:
+            raise ValueError("Invalid bounds for log-transformation: bounds must be positive")
+        if parameter.grid and min(parameter.grid) <= 0:
+            raise ValueError("Invalid grid values for log-transformation: values must be positive")
+
+    # parameter priors
+    if parameter.has_prior:
+        if not parameter.is_double:
+            raise ValueError("Prior only applies to parameters type of double")
+        if parameter.grid:
+            raise ValueError("Grid parameters cannot have priors")
+        if parameter.has_transformation:
+            raise ValueError("Parameters with log transformation cannot have priors")
+        if parameter.prior.is_normal:
+            if not parameter.bounds.is_value_within(parameter.prior.mean):
+                raise ValueError(f"parameter.prior.mean {parameter.prior.mean} must be within bounds {parameter.bounds}")
+        if not (parameter.prior.is_normal ^ parameter.prior.is_beta):
+            raise ValueError(f"{parameter.prior} must be either normal or beta")
+
+def validate_observation(observation, experiment):
+    for parameter in experiment.parameters:
+        if parameter_conditions_satisfied(parameter, observation.assignments):
+            observation_must_have_parameter(observation, parameter)
+        else:
+            observation_does_not_have_parameter(observation, parameter)
+
+    validate_observation_conditionals(observation, experiment.conditionals)
+
+    if experiment.is_multitask:
+        validate_observation_tasks(observation, experiment.tasks)
+
+    if not experiment.is_multitask and observation.task:
+        raise ValueError("Observation with task is not expected for this experiment")
+
+    if observation.failed and observation.metric_evaluations:
+        raise ValueError(
+          f"Observation marked as failure ({observation.failed}) should not have values. "
+          f"Observation metrics are: {observation.metric_evaluations}."
+        )
+
+    if not observation.failed:
+        num_reported_metrics = len(observation.metric_evaluations)
+        if num_reported_metrics != len(experiment.metrics):
+            raise ValueError("The number of observation values and experiment metrics must be equal.")
+        for m in experiment.metrics:
+            if observation.get_metric_evaluation_by_name(m.name) is None:
+                raise ValueError(
+                  f"Values must have metric names defined in experiment: {[m.name for m in experiment.metrics]}."
+                )
+
+
+def check_all_conditional_values_satisfied_for_experiment(experiment):
     num_conditional_values = numpy.product([len(c.values) for c in experiment.conditionals])
     satisfied_parameter_configurations = set([])
     for parameter in experiment.parameters:
@@ -199,11 +348,10 @@ def check_all_conditional_values_satisfied(experiment):
                 conditional_values.append([x.name for x in conditional.values])
         for selected_conditionals in itertools.product(*conditional_values):
             satisfied_parameter_configurations.add(selected_conditionals)
-
     if len(satisfied_parameter_configurations) != num_conditional_values:
         raise ValueError("Need at least one parameter that satisfies each conditional value")
 
-def check_constraint_feasibility(experiment):
+def check_constraint_feasibility_for_experiment(experiment):
     def parse_constraints_to_halfspaces(constraints, parameters):
         constrained_parameters_names = []
         for constraint in constraints:
@@ -243,7 +391,7 @@ def check_constraint_feasibility(experiment):
     if not feasibility:
         raise ValueError("Infeasible constraints")
 
-def validate_constraints(experiment):
+def validate_constraints_for_experiment(experiment):
     parameter_names = []
     double_params_names = []
     integer_params_names = []
@@ -301,7 +449,7 @@ def validate_constraints(experiment):
         if len(set(term_types)) > 1:
             raise ValueError("Constraint functions cannot mix integers and doubles. One or the other only.")
 
-def validate_tasks(experiment, class_name):
+def validate_tasks_for_experiment(experiment, class_name):
     if len(experiment.tasks) < 2:
         raise ValueError(f"For multitask {class_name}, at least 2 tasks must be present")
     costs = [t.cost for t in experiment.tasks]
@@ -313,3 +461,86 @@ def validate_tasks(experiment, class_name):
         raise ValueError(f"For multitask {class_name}, all task costs must be distinct")
     if 1 not in costs:
         raise ValueError(f"For multitask {class_name}, exactly one task must have cost == 1 (none present).")
+
+def validate_conditionals_for_experiment(experiment):
+    conditional_names = [c.name for c in experiment.conditionals]
+    if not len(conditional_names) == get_num_distinct_elements(conditional_names):
+        raise ValueError(f"No duplicate conditionals are allowed: {conditional_names}")
+
+    for parameter in experiment.parameters:
+        if parameter.conditions and any(c.name not in conditional_names for c in parameter.conditions):
+            unsatisfied_condition_names = [c.name for c in parameter.conditions if c.name not in conditional_names]
+            raise ValueError(
+              f"The parameter {parameter.name} has conditions {unsatisfied_condition_names} that are not part of"
+              " the conditionals"
+            )
+
+    check_all_conditional_values_satisfied_for_experiment(experiment)
+
+def validate_metrics_for_experiment(experiment):
+    metric_names = [m.name for m in experiment.metrics]
+    if not len(metric_names) == get_num_distinct_elements(metric_names):
+        raise ValueError(f"No duplicate metrics are allowed: {metric_names}")
+
+def validate_parameters_for_experiment(experiment):
+    param_names = [p.name for p in experiment.parameters]
+    if not len(param_names) == get_num_distinct_elements(param_names):
+        raise ValueError(f"No duplicate parameters are allowed: {param_names}")
+
+def observation_must_have_parameter(observation, parameter):
+    if parameter.name not in observation.assignments:
+        raise ValueError(
+          f"Parameter {parameter.name} is required for this experiment, "
+          f"and is missing from this observation: {observation.assignments}"
+        )
+    parameter_value = observation.assignments[parameter.name]
+    if parameter.is_categorical:
+        expected_categories = [cv.name for cv in parameter.categorical_values]
+        if parameter_value not in expected_categories:
+            raise ValueError(
+              f"Categorical parameter {parameter.name} must have one of following categories: "
+              f"{expected_categories} instead of {parameter_value}"
+            )
+    if parameter.grid and parameter_value not in parameter.grid:
+        raise ValueError(
+          f"Grid parameter {parameter.name} must have one of following grid values: "
+          f"{parameter.grid} instead of {parameter_value}"
+        )
+    if parameter.has_transformation:
+        if not (parameter_value > 0):
+            raise ValueError(f"Assignment must be positive for log-transformed parameter {parameter.name}")
+
+def observation_does_not_have_parameter(observation, parameter):
+    if parameter.name in observation.assignments:
+        raise ValueError(
+          f"Parameter {parameter.name} does not satisfy conditions. "
+          f"Observation assignments: {observation.assignments} is invalid."
+        )
+
+def validate_observation_conditionals(observation, conditionals):
+    for conditional in conditionals:
+        if conditional.name not in observation.assignments:
+            raise ValueError(f"Conditional parameter {conditional.name} must be in {observation}")
+        conditional_value = observation.assignments[conditional.name]
+        expected_conditional_options = [cv.name for cv in conditional.values]
+        if conditional_value not in expected_conditional_options:
+            raise ValueError(
+              f"Conditional parameter {conditional.name} must have one of following options: "
+              f"{expected_conditional_options} instead of {conditional_value}"
+            )
+
+def validate_observation_tasks(observation, tasks):
+    if not observation.task:
+        raise ValueError("Observation must have a task field for this experiment")
+    obs_task_name = observation.task.name
+    if obs_task_name not in [t.name for t in tasks]:
+        raise ValueError(
+          f"Task {obs_task_name} is not a valid task for this experiment. Must be one of the following: {tasks}"
+        )
+    obs_task_costs = observation.task.cost
+    expected_task_costs = [t.cost for t in tasks]
+    if obs_task_costs not in expected_task_costs:
+        raise ValueError(
+          f"Task cost {obs_task_costs} is not a valid cost for this experiment. Must be one of the following:"
+          f" {expected_task_costs}"
+        )
